useCheckBiometricAuthChanged
