+++ packages/kit/src/hooks/useBiometricAuthInfo.ts
@@ -1,12 +1,17 @@
-import { useMemo } from 'react';
+import { useCallback, useEffect, useMemo } from 'react';
 
 import { AuthenticationType } from 'expo-local-authentication';
 import { useIntl } from 'react-intl';
 
+import { onVisibilityStateChange } from '@onekeyhq/components';
 import type { IKeyOfIcons } from '@onekeyhq/components';
 import { usePasswordBiologyAuthInfoAtom } from '@onekeyhq/kit-bg/src/states/jotai/atoms';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
+import { checkBiometricAuthChanged } from '@onekeyhq/shared/src/modules3rdParty/check-biometric-auth-changed';
 import platformEnv from '@onekeyhq/shared/src/platformEnv';
+import biologyAuth from '@onekeyhq/shared/src/biologyAuth';
+
+import backgroundApiProxy from '../background/instance/backgroundApiProxy';
 
 export const useBiometricAuthInfo = () => {
   const [{ authType }] = usePasswordBiologyAuthInfoAtom();
@@ -42,3 +47,35 @@ export const useBiometricAuthInfo = () => {
     };
   }, [authType, intl]);
 };
+
+export const useCheckBiometricAuthChanged =
+  platformEnv.isNative || platformEnv.isDesktopMac
+    ? () => {
+        const checkAuthChanged = useCallback(async () => {
+          const isSupport = await biologyAuth.isSupportBiologyAuth();
+          if (!isSupport) {
+            return;
+          }
+          const changed = await checkBiometricAuthChanged();
+          if (changed) {
+            void backgroundApiProxy.servicePassword.setBiologyAuthEnable(false);
+            alert(
+              'Biometric authentication has changed. Please manually re-enable it.',
+            );
+          }
+        }, []);
+
+        useEffect(() => {
+          void checkAuthChanged();
+          const handleVisibilityStateChange = (visible: boolean) => {
+            if (visible) {
+              void checkAuthChanged();
+            }
+          };
+          const removeSubscription = onVisibilityStateChange(
+            handleVisibilityStateChange,
+          );
+          return removeSubscription;
+        }, [checkAuthChanged]);
+      }
+    : () => {};
